// src/lib/enums.generated.ts
export const LoanTypeOptions = [
  { value: "dscr", label: "Dscr" },
  { value: "fix_and_flip", label: "Fix And Flip" },
  { value: "ground_up", label: "Ground Up" },
  { value: "bridge", label: "Bridge" },
  { value: "portfolio", label: "Portfolio" },
  { value: "cash_out_refi", label: "Cash Out Refi" }
] as const;
// ...
export const PropertyTypeOptions = [
  { value: "single_family", label: "Single Family" },
  { value: "2_4_units", label: "2 4 Units" },
  { value: "5_8_units", label: "5 8 Units" },
  { value: "mixed_use", label: "Mixed Use" },
  { value: "commercial", label: "Commercial" }
] as const;